+++ src/components/Validator.js
@@ -3,17 +3,17 @@ import _each from 'lodash/each';
 import _has from 'lodash/has';
 import _isNumber from 'lodash/isNumber';
 import FormioUtils from '../utils/index';
-export var Validator = {
+export const Validator = {
   get: _get,
   each: _each,
   has: _has,
-  checkValidator: function(component, validator, setting, value, data) {
+  checkValidator(component, validator, setting, value, data) {
     // Make sure this component isn't conditionally disabled.
     if (!FormioUtils.checkCondition(component.component, data, component.data)) {
       return '';
     }
 
-    let result = validator.check.call(this, component, setting, value, data);
+    const result = validator.check.call(this, component, setting, value, data);
     if (typeof result === 'string') {
       return result;
     }
@@ -22,20 +22,20 @@ export var Validator = {
     }
     return '';
   },
-  validate: function(component, validator, value, data) {
+  validate(component, validator, value, data) {
     if (validator.key && _has(component.component, validator.key)) {
-      let setting = this.get(component.component, validator.key);
+      const setting = this.get(component.component, validator.key);
       return this.checkValidator(component, validator, setting, value, data);
     }
     return this.checkValidator(component, validator, null, value, data);
   },
-  check: function(component, data) {
+  check(component, data) {
     let result = '';
-    let value = component.getRawValue();
+    const value = component.getRawValue();
     data = data || component.data;
     _each(component.validators, (name) => {
       if (this.validators.hasOwnProperty(name)) {
-        let validator = this.validators[name];
+        const validator = this.validators[name];
         if (component.validateMultiple(value)) {
           _each(value, (val) => {
             result = this.validate(component, validator, val, data);
@@ -52,15 +52,21 @@ export var Validator = {
         }
       }
     });
+
+    const customErrorMessage = _get(component, 'component.validate.customMessage');
+    if (result && customErrorMessage) {
+      result = component.t(customErrorMessage);
+    }
+
     return result;
   },
   validators: {
     required: {
       key: 'validate.required',
-      message: function(component, setting) {
+      message(component, setting) {
         return component.t(component.errorMessage('required'), {field: component.errorLabel});
       },
-      check: function(component, setting, value) {
+      check(component, setting, value) {
         if (!FormioUtils.boolValue(setting)) {
           return true;
         }
@@ -69,14 +75,14 @@ export var Validator = {
     },
     min: {
       key: 'validate.min',
-      message: function(component, setting) {
+      message(component, setting) {
         return component.t(component.errorMessage('min'), {
           field: component.errorLabel,
           min: parseFloat(setting)
         });
       },
-      check: function(component, setting, value) {
-        let min = parseFloat(setting);
+      check(component, setting, value) {
+        const min = parseFloat(setting);
         if (!min || (!_isNumber(value))) {
           return true;
         }
@@ -85,14 +91,14 @@ export var Validator = {
     },
     max: {
       key: 'validate.max',
-      message: function(component, setting) {
+      message(component, setting) {
         return component.t(component.errorMessage('max'), {
           field: component.errorLabel,
           max: parseFloat(setting)
         });
       },
-      check: function(component, setting, value) {
-        let max = parseFloat(setting);
+      check(component, setting, value) {
+        const max = parseFloat(setting);
         if (!max || (!_isNumber(value))) {
           return true;
         }
@@ -101,14 +107,14 @@ export var Validator = {
     },
     minLength: {
       key: 'validate.minLength',
-      message: function(component, setting) {
+      message(component, setting) {
         return component.t(component.errorMessage('minLength'), {
           field: component.errorLabel,
           length: (setting - 1)
         });
       },
-      check: function(component, setting, value) {
-        let minLength = parseInt(setting, 10);
+      check(component, setting, value) {
+        const minLength = parseInt(setting, 10);
         if (!minLength || (typeof value !== 'string')) {
           return true;
         }
@@ -117,14 +123,14 @@ export var Validator = {
     },
     maxLength: {
       key: 'validate.maxLength',
-      message: function(component, setting) {
+      message(component, setting) {
         return component.t(component.errorMessage('maxLength'), {
           field: component.errorLabel,
           length: (setting + 1)
         });
       },
-      check: function(component, setting, value) {
-        let maxLength = parseInt(setting, 10);
+      check(component, setting, value) {
+        const maxLength = parseInt(setting, 10);
         if (!maxLength || (typeof value !== 'string')) {
           return true;
         }
@@ -132,50 +138,50 @@ export var Validator = {
       }
     },
     email: {
-      message: function(component, setting) {
+      message(component, setting) {
         return component.t(component.errorMessage('invalid_email'), {
           field: component.errorLabel
         });
       },
-      check: function(component, setting, value) {
+      check(component, setting, value) {
         // From http://stackoverflow.com/questions/46155/validate-email-address-in-javascript
-        var re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+        const re = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
 
         // Allow emails to be valid if the component is pristine and no value is provided.
         return (component.pristine && !value) || re.test(value);
       }
     },
     date: {
-      message: function(component, setting) {
+      message(component, setting) {
         return component.t(component.errorMessage('invalid_date'), {
           field: component.errorLabel
         });
       },
-      check: function(component, setting, value) {
+      check(component, setting, value) {
         return (value !== 'Invalid date');
       }
     },
     pattern: {
       key: 'validate.pattern',
-      message: function(component, setting) {
-        return component.t(component.errorMessage('pattern'), {
+      message(component, setting) {
+        return component.t(_get(component, 'component.validate.patternMessage', component.errorMessage('pattern'), {
           field: component.errorLabel,
           pattern: setting
-        });
+        }));
       },
-      check: function(component, setting, value) {
-        let pattern = setting;
+      check(component, setting, value) {
+        const pattern = setting;
         if (!pattern) {
           return true;
         }
-        let regexStr = '^' + pattern + '$';
-        let regex = new RegExp(regexStr);
+        const regexStr = '^' + pattern + '$';
+        const regex = new RegExp(regexStr);
         return regex.test(value);
       }
     },
     json: {
       key: 'validate.json',
-      check: function(component, setting, value, data) {
+      check(component, setting, value, data) {
         if (!setting) {
           return true;
         }
@@ -194,12 +200,12 @@ export var Validator = {
     },
     custom: {
       key: 'validate.custom',
-      message: function(component) {
+      message(component) {
         return component.t(component.errorMessage('custom'), {
           field: component.errorLabel
         });
       },
-      check: function(component, setting, value, data) {
+      check(component, setting, value, data) {
         if (!setting) {
           return true;
         }
@@ -209,7 +215,7 @@ export var Validator = {
         /*eslint-disable no-unused-vars */
         var input = value;
         /*eslint-enable no-unused-vars */
-        custom = custom.replace(/({{\s+(.*)\s+}})/, function (match, $1, $2) {
+        custom = custom.replace(/({{\s+(.*)\s+}})/, (match, $1, $2) => {
           if ($2.indexOf('data.') === 0) {
             return _get(data, $2.replace('data.', ''));
           }